Alternative formula constructions
Sometimes %{cf_id} construction don't works because we don't have
control over the field contents. A custom field can be empty or
even a value that lead us to invalid Ruby syntax.

The solution here just make available a local cfs array with custom
fields values. So, instead of %{cf_1} we can use cfs[1], it is very
useful when we have things like "%{cf_1} * %{cf_2}" that can be
translated to " * %{cf_2}" and trow a Ruby systax error, now we can
just user "cfs[1] * cfs[2]" or even "cfs[1].to_f * cfs[2].to_f"
without worry about empty values.

Note that old behavior still works and is used by default when we
double-click on an item in the list of available fields.